Add skip-cleanup, reuse-db, port flags

+++ components/director/run.sh
@@ -11,6 +11,9 @@ set -e
 
 ROOT_PATH=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 
+SKIP_DB_CLEANUP=false
+REUSE_DB=false
+
 POSITIONAL=()
 while [[ $# -gt 0 ]]
 do
@@ -22,8 +25,22 @@ do
             SKIP_APP_START=true
             shift # past argument
         ;;
+        --skip-db-cleanup)
+            SKIP_DB_CLEANUP=true
+            shift
+        ;;
+        --reuse-db)
+            REUSE_DB=true
+            shift
+        ;;
         --debug)
             DEBUG=true
+            DEBUG_PORT=40000
+            shift
+        ;;
+        --debug-port)
+            DEBUG_PORT=$2
+            shift
             shift
         ;;
         --*)
@@ -33,7 +50,6 @@ do
     esac
 done
 set -- "${POSITIONAL[@]}" # restore positional parameters
-
 
 POSTGRES_CONTAINER="test-postgres"
 POSTGRES_VERSION="11"
@@ -45,48 +61,62 @@ DB_PORT="5432"
 DB_HOST="127.0.0.1"
 
 function cleanup() {
-    echo -e "${GREEN}Cleanup Postgres container${NC}"
-    docker rm --force ${POSTGRES_CONTAINER}
+    if [[ ${DEBUG} ]]; then
+       echo -e "${GREEN}Cleanup Director binary${NC}"
+       rm  $GOPATH/src/github.com/kyma-incubator/compass/components/director/director
+    fi
+
+    if [[ ${SKIP_DB_CLEANUP} = false ]]; then
+        echo -e "${GREEN}Cleanup Postgres container${NC}"
+        docker rm --force ${POSTGRES_CONTAINER}
+    else
+        echo -e "${GREEN}Skipping Postgres container cleanup${NC}"
+    fi
 }
 
 trap cleanup EXIT
 
-echo -e "${GREEN}Start Postgres in detached mode${NC}"
-docker run -d --name ${POSTGRES_CONTAINER} \
-            -e POSTGRES_HOST=${DB_HOST} \
-            -e POSTGRES_USER=${DB_USER} \
-            -e POSTGRES_PASSWORD=${DB_PWD} \
-            -e POSTGRES_DB=${DB_NAME} \
-            -e POSTGRES_PORT=${DB_PORT} \
-            -p ${DB_PORT}:${DB_PORT} \
-            postgres:${POSTGRES_VERSION}
 
-set +e
-echo '# WAITING FOR CONNECTION WITH DATABASE #'
-for i in {1..30}
-do
-    docker exec ${POSTGRES_CONTAINER} pg_isready -U "${DB_USER}" -h "${DB_HOST}" -p "${DB_PORT}" -d "${DB_NAME}"
-    if [ $? -eq 0 ]
-    then
-        dbReady=true
-        break
+if [[ ${REUSE_DB} = true ]]; then
+    echo -e "${GREEN}Will reuse existing Postgres container${NC}"
+else
+    echo -e "${GREEN}Start Postgres in detached mode${NC}"
+    docker run -d --name ${POSTGRES_CONTAINER} \
+                -e POSTGRES_HOST=${DB_HOST} \
+                -e POSTGRES_USER=${DB_USER} \
+                -e POSTGRES_PASSWORD=${DB_PWD} \
+                -e POSTGRES_DB=${DB_NAME} \
+                -e POSTGRES_PORT=${DB_PORT} \
+                -p ${DB_PORT}:${DB_PORT} \
+                postgres:${POSTGRES_VERSION}
+
+    set +e
+    echo '# WAITING FOR CONNECTION WITH DATABASE #'
+    for i in {1..30}
+    do
+        docker exec ${POSTGRES_CONTAINER} pg_isready -U "${DB_USER}" -h "${DB_HOST}" -p "${DB_PORT}" -d "${DB_NAME}"
+        if [ $? -eq 0 ]
+        then
+            dbReady=true
+            break
+        fi
+        sleep 1
+    done
+
+    if [ "${dbReady}" != true ] ; then
+        echo '# COULD NOT ESTABLISH CONNECTION TO DATABASE #'
+        exit 1
     fi
-    sleep 1
-done
 
-if [ "${dbReady}" != true ] ; then
-    echo '# COULD NOT ESTABLISH CONNECTION TO DATABASE #'
-    exit 1
+    set -e
+
+    echo -e "${GREEN}Populate DB${NC}"
+
+    cat ${ROOT_PATH}/../schema-migrator/migrations/director/*.up.sql | \
+        docker exec -i ${POSTGRES_CONTAINER} psql -U "${DB_USER}" -h "${DB_HOST}" -p "${DB_PORT}" -d "${DB_NAME}"
+    cat ${ROOT_PATH}/../schema-migrator/seeds/director/*.sql | \
+        docker exec -i ${POSTGRES_CONTAINER} psql -U "${DB_USER}" -h "${DB_HOST}" -p "${DB_PORT}" -d "${DB_NAME}"
 fi
-
-set -e
-
-echo -e "${GREEN}Populate DB${NC}"
-
-cat ${ROOT_PATH}/../schema-migrator/migrations/director/*.up.sql | \
-    docker exec -i ${POSTGRES_CONTAINER} psql -U "${DB_USER}" -h "${DB_HOST}" -p "${DB_PORT}" -d "${DB_NAME}"
-cat ${ROOT_PATH}/../schema-migrator/seeds/director/*.sql | \
-    docker exec -i ${POSTGRES_CONTAINER} psql -U "${DB_USER}" -h "${DB_HOST}" -p "${DB_PORT}" -d "${DB_NAME}"
 
 if [[  ${SKIP_APP_START} ]]; then
     echo -e "${GREEN}Skipping starting application${NC}"
@@ -111,12 +141,10 @@ export APP_CONNECTOR_URL="http://connector.not.configured.url/graphql"
 export APP_LEGACY_CONNECTOR_URL="https://adapter-gateway.kyma.local/v1/applications/signingRequests/info"
 
 if [[  ${DEBUG} ]]; then
-    PORT=40000
-    echo -e "${GREEN}Debug mode activated on port $PORT${NC}"
-    cd $GOPATH/src/github.com/kyma-incubator/compass/components/director && \
-    CGO_ENABLED=0 go build -gcflags="all=-N -l" ./cmd/director && \
-    dlv --listen=:$PORT --headless=true --api-version=2 exec ./director
-    rm ./director
+    echo -e "${GREEN}Debug mode activated on port $DEBUG_PORT${NC}"
+    cd $GOPATH/src/github.com/kyma-incubator/compass/components/director
+    CGO_ENABLED=0 go build -gcflags="all=-N -l" ./cmd/director
+    dlv --listen=:$DEBUG_PORT --headless=true --api-version=2 exec ./director
 else
     go run ${ROOT_PATH}/cmd/director/main.go
 fi
